+++ CHANGES.md
@@ -1,3 +1,7 @@
+## Next
+
+- [#3](https://github.com/dmolesUC3/typesafe_enum/pull/3) - No need for `instance_eval` when creating new enum instance methods - [@dblock](https://github.com/dblock).
+
 ## 0.1.2
 
 - Fixed issue where `::find_by_value_str` failed to return `nil` for bad values